chore: move dummy quote to test folder
