+++ test/data/bridge/dummy-quotes.ts
